+++ tagpr.go
@@ -6,7 +6,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"log"
 	"net/url"
 	"os"
 	"path/filepath"
@@ -14,7 +13,6 @@ import (
 	"slices"
 	"strconv"
 	"strings"
-	"text/template"
 	"time"
 
 	"github.com/Songmu/gitconfig"
@@ -341,7 +339,11 @@ func (tp *tagpr) isTagPR(pr *github.PullRequest) bool {
 	return false
 }
 
-func (tp *tagpr) Run(ctx context.Context) error {
+func (tp *tagpr) Run(ctx context.Context) (runErr error) {
+	var commandErr error
+	defer func() {
+		runErr = errors.Join(runErr, commandErr)
+	}()
 	if err := tp.runOptions.validate(); err != nil {
 		return err
 	}
@@ -495,7 +497,7 @@ func (tp *tagpr) Run(ctx context.Context) error {
 	}
 
 	if prog := tp.cfg.Command(); prog != "" {
-		tp.Exec(prog, currVer, nextVer)
+		tp.runReleaseCommand(&commandErr, configCommand, prog, currVer, nextVer)
 	}
 
 	if len(vfiles) > 0 && vfiles[0] != "" {
@@ -508,7 +510,7 @@ func (tp *tagpr) Run(ctx context.Context) error {
 	tp.c.Git("add", "-f", tp.cfg.conf) // ignore any errors
 
 	if prog := tp.cfg.PostVersionCommand(); prog != "" {
-		tp.Exec(prog, currVer, nextVer)
+		tp.runReleaseCommand(&commandErr, configPostVersionCommand, prog, currVer, nextVer)
 	}
 
 	releaseYaml := tp.cfg.ReleaseYAMLPath()
@@ -772,44 +774,76 @@ func (tp *tagpr) Run(ctx context.Context) error {
 		}
 	}
 
-	opts := []gh2changelog.Option{
-		gh2changelog.GitPath(tp.gitPath),
-		gh2changelog.SetOutputs(tp.c.outStream, tp.c.errStream),
-		gh2changelog.GitHubClient(tp.gh),
-		gh2changelog.TagPrefix(tp.normalizedTagPrefix),
-		gh2changelog.ChangelogMdPath(tp.cfg.ChangelogFile()),
+	host := "github.com"
+	if tp.gh.BaseURL != nil {
+		host = strings.TrimPrefix(tp.gh.BaseURL.Host, "api.")
 	}
-	if tp.cfg.ReleaseYAMLPath() != "" {
-		opts = append(opts, gh2changelog.ReleaseYamlPath(tp.cfg.ReleaseYAMLPath()))
-	}
-	if fixedMajor, err := tp.cfg.FixedMajorVersion(); err == nil && fixedMajor != nil {
-		opts = append(opts, gh2changelog.FilteredMajorVersion(*fixedMajor))
-	}
-	if tp.cfg.CalendarVersioning() && latestSemverTag != "" {
-		opts = append(opts, gh2changelog.VersionTags([]string{latestSemverTag}))
-	}
-	gch, err := gh2changelog.New(ctx, opts...)
-	if err != nil {
-		return err
-	}
+	currTag := fullTag(tp.normalizedTagPrefix, currVer.Tag())
+	nextTag := fullTag(tp.normalizedTagPrefix, nextVer.Tag())
+	draft := &lazyDraftChangelog{load: func() (*draftChangelog, error) {
+		opts := []gh2changelog.Option{
+			gh2changelog.GitPath(tp.gitPath),
+			gh2changelog.SetOutputs(tp.c.outStream, tp.c.errStream),
+			gh2changelog.GitHubClient(tp.gh),
+			gh2changelog.TagPrefix(tp.normalizedTagPrefix),
+			gh2changelog.ChangelogMdPath(tp.cfg.ChangelogFile()),
+		}
+		if tp.cfg.ReleaseYAMLPath() != "" {
+			opts = append(opts, gh2changelog.ReleaseYamlPath(tp.cfg.ReleaseYAMLPath()))
+		}
+		if fixedMajor, err := tp.cfg.FixedMajorVersion(); err == nil && fixedMajor != nil {
+			opts = append(opts, gh2changelog.FilteredMajorVersion(*fixedMajor))
+		}
+		if tp.cfg.CalendarVersioning() && latestSemverTag != "" {
+			opts = append(opts, gh2changelog.VersionTags([]string{latestSemverTag}))
+		}
+		gch, err := gh2changelog.New(ctx, opts...)
+		if err != nil {
+			return nil, err
+		}
 
-	draftNextTag := fullTag(tp.normalizedTagPrefix, nextVer.Tag())
-	changelog, orig, err := gch.Draft(ctx, draftNextTag, releaseBranch, time.Now())
-	if err != nil {
-		return err
+		draftNextTag := fullTag(tp.normalizedTagPrefix, nextVer.Tag())
+		changelog, generatedNotes, err := gch.Draft(ctx, draftNextTag, releaseBranch, time.Now())
+		if err != nil {
+			return nil, err
+		}
+		return &draftChangelog{
+			generator:      gch,
+			changelog:      changelog,
+			generatedNotes: generatedNotes,
+		}, nil
+	}}
+	prArg := &tmplArg{
+		NextVersion: nextVer.Tag(),
+		Branch:      rcBranch,
+		TagPrefix:   strings.TrimSuffix(tp.normalizedTagPrefix, "/"),
+		loadChangelog: func() (string, error) {
+			generatedNotes, err := draft.GeneratedNotes()
+			if err != nil {
+				return "", err
+			}
+			return replaceCompareLink(
+				generatedNotes, host, tp.owner, tp.repo, currTag, nextTag, rcBranch), nil
+		},
 	}
+	pt := loadPRTmpl(tp.cfg)
 
-	if tp.cfg.changelog == nil || *tp.cfg.changelog {
+	if tp.cfg.Changelog() {
+		result, err := draft.Load()
+		if err != nil {
+			return err
+		}
 		changelogMd := tp.cfg.ChangelogFile()
+		changelog := result.changelog
 		if !exists(changelogMd) {
-			logs, _, err := gch.Changelogs(ctx, 20)
+			logs, _, err := result.generator.Changelogs(ctx, 20)
 			if err != nil {
 				return err
 			}
 			changelog = strings.Join(
 				append([]string{changelog}, logs...), "\n")
 		}
-		if _, err := gch.Update(changelog, 0); err != nil {
+		if _, err := result.generator.Update(changelog, 0); err != nil {
 			return err
 		}
 
@@ -843,6 +877,11 @@ func (tp *tagpr) Run(ctx context.Context) error {
 		}
 	}
 
+	prText, err := pt.Render(prArg)
+	if err != nil {
+		return err
+	}
+
 	// Create or Get remote rcBranch reference
 	rcBranchRef := "refs/heads/" + rcBranch
 	_, resp, err = tp.gh.Git.GetRef(ctx, tp.owner, tp.repo, rcBranchRef)
@@ -868,41 +907,6 @@ func (tp *tagpr) Run(ctx context.Context) error {
 	_, resp, err = tp.gh.Git.UpdateRef(ctx, tp.owner, tp.repo, rcBranchRef, updateRef)
 	if err != nil {
 		showGHError(err, resp)
-		return err
-	}
-
-	var tmpl *template.Template
-	if t := tp.cfg.Template(); t != "" {
-		tmpTmpl, err := template.ParseFiles(t)
-		if err == nil {
-			tmpl = tmpTmpl
-		} else {
-			log.Printf("parse configured template failed: %s\n", err)
-		}
-	} else if t := tp.cfg.TemplateText(); t != "" {
-		tmpTmplTxt, err := template.New("templateText").Parse(t)
-		if err == nil {
-			tmpl = tmpTmplTxt
-		} else {
-			log.Printf("parse configured template failed: %s\n", err)
-		}
-	}
-
-	host := "github.com"
-	if tp.gh.BaseURL != nil {
-		host = strings.TrimPrefix(tp.gh.BaseURL.Host, "api.")
-	}
-	currTag := fullTag(tp.normalizedTagPrefix, currVer.Tag())
-	nextTag := fullTag(tp.normalizedTagPrefix, nextVer.Tag())
-	orig = replaceCompareLink(orig, host, tp.owner, tp.repo, currTag, nextTag, rcBranch)
-	pt := newPRTmpl(tmpl)
-	prText, err := pt.Render(&tmplArg{
-		NextVersion: nextVer.Tag(),
-		Branch:      rcBranch,
-		Changelog:   orig,
-		TagPrefix:   strings.TrimSuffix(tp.normalizedTagPrefix, "/"),
-	})
-	if err != nil {
 		return err
 	}
 
@@ -1023,17 +1027,38 @@ func (tp *tagpr) getVfiles(currVer *semv) ([]string, error) {
 	return vfiles, nil
 }
 
-func (tp *tagpr) Exec(prog string, currVer, nextVer *semv) {
+func (tp *tagpr) Exec(name, command string, currVer, nextVer *semv) error {
+	prog := command
 	var progArgs []string
 	if strings.ContainsAny(prog, " \n") {
 		progArgs = []string{"-c", prog}
 		prog = "sh"
 	}
-	tp.c.Cmd(prog, progArgs, map[string]string{
+	_, stderr, err := tp.c.Cmd(prog, progArgs, map[string]string{
 		"TAGPR_CURRENT_VERSION": currVer.Tag(),
 		"TAGPR_NEXT_VERSION":    nextVer.Tag(),
 	})
+	if err == nil {
+		return nil
+	}
 
+	commandErr := fmt.Errorf("%s %q failed: %w", name, command, err)
+	if stderr != "" {
+		commandErr = fmt.Errorf("%w\n%s", commandErr, stderr)
+	}
+	return commandErr
+}
+
+func (tp *tagpr) showReleaseCommandError(name string, err error) {
+	fmt.Fprintln(tp.c.errStream, err)
+	showErrorAnnotation(tp.c.errStream, name+" failed", err.Error())
+}
+
+func (tp *tagpr) runReleaseCommand(errs *error, name, command string, currVer, nextVer *semv) {
+	if err := tp.Exec(name, command, currVer, nextVer); err != nil {
+		*errs = errors.Join(*errs, err)
+		tp.showReleaseCommandError(name, err)
+	}
 }
 
 func (tp *tagpr) defaultBranch() (string, error) {